+++ recipes/default.rb
@@ -17,4 +17,4 @@
 # limitations under the License.
 #
 
-include_recipe "ipxe::build"
+include_recipe "ipxe::install"